backend/requirements.txt
fastapi
supabase
python-dotenv
imapclient
uvicorn
pydantic
gunicorn
jwt